requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
openai>=1.0.0
python-dotenv>=0.19.0
chromadb
langchain
langchain-community
langchain-openai
django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
pymysql
passlib[bcrypt]
python-jose[cryptography]
